requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.2.2
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.7
django-channels==0.7.0
django-cors-headers==3.7.0
djangochannelsrestframework==0.2.1
djangorestframework==3.12.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==21.2.1
urllib3==1.26.3
zope.interface==5.2.0
